+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.5.0
+- Updated minimum required Okteto CLI version to 3.0.0
+- Added a command to run tests using "Okteto Test"
+- Removed "okteto create" command since the CLI doesn't support it anymore
+- Removed support for Okteto Stacks or Okteto Manifest 1.0 since they have been fully deprecated
+- Namespace is no longer read from the manifest; instead, you need to set it via the `context` command
+- Updated dependencies
+
 ## 0.4.15
 - Updated minimum required Okteto CLI version to 2.29.3
 - Updated dependencies